+++ app/src/main/res/layout/activity_filter.xml
@@ -10,7 +10,7 @@
     <androidx.constraintlayout.widget.ConstraintLayout
         android:layout_width="match_parent"
         android:layout_height="match_parent"
-        tools:context="com.seoulfitu.android.ui.filter.FilterActivity">
+        tools:context="com.seoulfitu.android.ui.facilityfilter.FacilityFilterActivity">
 
         <ScrollView
             android:layout_width="match_parent"
@@ -25,7 +25,7 @@
                 android:layout_height="wrap_content"
                 android:orientation="vertical">
 
-                <com.seoulfitu.android.ui.filter.FilterCustomView
+                <com.seoulfitu.android.ui.common.custom.FilterCustomView
                     android:id="@+id/cv_filter_city"
                     android:layout_width="match_parent"
                     android:layout_height="wrap_content"
@@ -33,7 +33,7 @@
                     android:layout_marginTop="24dp"
                     android:layout_marginEnd="24dp"/>
 
-                <com.seoulfitu.android.ui.filter.FilterCustomView
+                <com.seoulfitu.android.ui.common.custom.FilterCustomView
                     android:id="@+id/cv_filter_facility"
                     android:layout_width="match_parent"
                     android:layout_height="wrap_content"
@@ -41,7 +41,7 @@
                     android:layout_marginTop="24dp"
                     android:layout_marginEnd="24dp" />
 
-                <com.seoulfitu.android.ui.filter.FilterCustomView
+                <com.seoulfitu.android.ui.common.custom.FilterCustomView
                     android:id="@+id/cv_filter_rent"
                     android:layout_width="match_parent"
                     android:layout_height="wrap_content"
@@ -49,7 +49,7 @@
                     android:layout_marginTop="24dp"
                     android:layout_marginEnd="24dp" />
 
-                <com.seoulfitu.android.ui.filter.FilterCustomView
+                <com.seoulfitu.android.ui.common.custom.FilterCustomView
                     android:id="@+id/cv_filter_parking"
                     android:layout_width="match_parent"
                     android:layout_height="wrap_content"
